requirements.txt
gcsfs
google-cloud-bigquery
pandas